refactor: name file jsx
